Improve configuration of DefaultLinker
Enable PROPERTY_FILTER_IMPORTDECLARATION and PROPERTY_FILTER_IMPORTERSERVICE as
mandatory properties. Theses properties are now used by the DefaultLinker.
Use FuchsiaUtils.getFilter(..) for the treatments of the filter properties.

Add the PROPERTY_UNIQUE_IMPORTATION property which is a boolean. This property
force the DefaultLinker to give an ImportDeclaration to at most one
ImporterService. This property is set to false by default.
WARNING : this property is actually based on the number of ImporterService
actually bind to the ImportDeclaration, i.e. if an other Linker has already bind
the ImportDeclaration to an ImporterService, the DefaultLinker will not give the
ImportDeclaration to any of its ImporterService.

Update the FoolInitializer to provides the 2 mandatory properties to the
DefaultLinker.