Change in declaration to provide ExportDeclarationi
The ImportDeclaration* are renamed Declaration*.
Declarations are now bind to ServiceReference (ImportDeclaration were bind to
ImporterService before that), Status is updated in that way.
ImportDeclaration and ExportDeclaration interfaces extends Declaration and
share their implementation.
ImportDeclarationBuilder and ExportDeclarationBuilder extend DeclarationBuilder.
They are functionnaly equivalent (provide the same methods) but permit to give
to the developer of DiscoveryService or ExportManager (for future) a explicitly
specialized builder.

It's a preliminary work to begin the implementation of the "export-chain" in
Fuchsia.