Add the full exportation chain.
ExportManagerService is the interface that must be provided by component
publishing ExportDeclaration. AbstractExportManagerComponent give a base to
implement such components.

ExportationLinker is the interface which must be provides by components used to
make the link between ExportDeclaration and ExporterService.
DefaultExportationLinker is the default implementation of the ExportationLinker
interface which is provided by Fuchsia.

ExporterService is the interface which must be provided by the component that
reify ExportDeclaration into something. AbstractExporterComponent give a base to
implement such components.