Update Fake module to handle more dynamicity
Update the FakeDiscoveryBridge to use file describing devices to create
on the fly ImportDeclaration. Devices are described through cfg files.
The CustomDeviceFileDeployer is used to monitor specials directories
and create the ImportDeclaration based on the cfg file.
Dynamic add and remove are supported.
In the same time that a ImportDeclaration is published, we create an fake device
on the platform through the Everest client

ImporterService has been updated to be compliant with the new discovery.